requirements.txt
Flask
pandas
torch
transformers
Pillow
